# Pull lazily on read for read-your-writes