#2105

Before:
```cmd
$ hyperfine target/debug/libsql-probe
Benchmark 1: target/debug/libsql-probe
  Time (mean ± σ):     11.488 s ±  3.327 s    [User: 0.367 s, System: 0.111 s]
  Range (min … max):    6.270 s … 17.406 s    10 runs
```

After:
```cmd
$ hyperfine target/debug/libsql-probe
Benchmark 1: target/debug/libsql-probe
  Time (mean ± σ):      3.156 s ±  1.063 s    [User: 0.406 s, System: 0.066 s]
  Range (min … max):    1.764 s …  5.268 s    10 runs
```

On this kind of workload, running the server locally:

```rs
for i in 0..100 {
    conn.execute("insert into test values (?)", libsql::params!(i)).await?;
}

let mut rows = conn.query("select * from test", ()).await?;
```